pyproject.toml
[tool.poetry]
name = "injecterr"
version = "0.1.0"
description = "A simple python framework to inject errors while running HiBench workloads on Hadoop and collecting logs"
authors = ["Gunwant Jain <mail@wantguns.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
fabric = "^2.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
injecterr = 'injecterr.main:main'

